+++ api/lib/notifications/notifications.js
@@ -3,17 +3,17 @@ const Sentry = require('@sentry/node')
 const {
   eventSubscription,
   publishNotificationAdded
-} = require('./subscriptions')
+} = require('../subscriptions')
 const {
   eventTypes,
   resourceTypes,
   getDefaultSubscriptions
 } = require('./notifications-types')
-const database = require('./database')
-const config = require('../config.js')
+const database = require('../database')
+const config = require('../../config.js')
 
 function getMessageTitle(networks, notification) {
-  const data = JSON.parse(notification.data)
+  const data = notification.properties || JSON.parse(notification.data)
   switch (notification.eventType) {
     case eventTypes.TRANSACTION_RECEIVE:
       return `You have received ${data.details.amount.amount} ${
@@ -53,9 +53,9 @@ function getMessageTitle(networks, notification) {
 
       return `Voting power increased from ${Number(
         data.prevValidator.votingPower
-      ).toFixed(6)} to ${Number(data.nextValidator.votingPower).toFixed(
-        6
-      )} (${percentageDifference.toFixed(3)}% increase) for ${
+      ).toFixed(2)}% to ${Number(data.nextValidator.votingPower).toFixed(
+        2
+      )}% (${percentageDifference.toFixed(3)}% increase) for ${
         data.nextValidator.name
       }`
     }
@@ -69,9 +69,9 @@ function getMessageTitle(networks, notification) {
 
       return `Voting power decreased from ${Number(
         data.prevValidator.votingPower
-      ).toFixed(6)} to ${Number(data.nextValidator.votingPower).toFixed(
-        6
-      )} (${percentageDifference.toFixed(3)}% decrease) for ${
+      ).toFixed(2)}% to ${Number(data.nextValidator.votingPower).toFixed(
+        2
+      )}% (${percentageDifference.toFixed(2)}% decrease) for ${
         data.nextValidator.name
       }`
     }
@@ -121,11 +121,11 @@ function findNetworkSlug(networks, networkId) {
 
 function getPushLink(
   networks,
-  { resourceType, eventType, networkId, resourceId, data }
+  { resourceType, eventType, networkId, resourceId, data, properties }
 ) {
   const resource =
     resourceType === resourceTypes.VALIDATOR ? eventType : resourceType
-  const notificationData = JSON.parse(data)
+  const notificationData = properties || JSON.parse(data)
 
   switch (resource) {
     case resourceTypes.TRANSACTION:
@@ -156,8 +156,8 @@ function getPushLink(
 
 // Get relevant icon for notification
 // TODO: Upload icons to DO instead of passing relative links
-function getIcon({ eventType, data }) {
-  const notificationData = JSON.parse(data)
+function getIcon({ eventType, data, properties }) {
+  const notificationData = properties || JSON.parse(data)
   switch (eventType) {
     case eventTypes.TRANSACTION_RECEIVE:
       return `/img/icons/activity/Received.svg`
@@ -166,6 +166,8 @@ function getIcon({ eventType, data }) {
     case eventTypes.PROPOSAL_CREATE:
     case eventTypes.PROPOSAL_UPDATE:
       return `/img/icons/activity/Submitted.svg`
+    case eventTypes.VALIDATOR_ADDED:
+      return `/img/networks/${notificationData.nextValidator.networkId}.png`
     case eventTypes.VALIDATOR_WEBSITE:
     case eventTypes.VALIDATOR_COMMISSION:
     case eventTypes.VALIDATOR_VOTING_POWER_INCREASE:
@@ -174,7 +176,6 @@ function getIcon({ eventType, data }) {
     case eventTypes.VALIDATOR_PICTURE:
     case eventTypes.VALIDATOR_STATUS:
     case eventTypes.VALIDATOR_MAX_CHANGE_COMMISSION:
-    case eventTypes.VALIDATOR_ADDED:
     case eventTypes.SLASH:
     case eventTypes.LIVENESS:
       // Picture field for validator type can be null
@@ -274,14 +275,16 @@ const getNotifications = (networks) => async (
     timestamp
   )
 
-  const notifications = relevantNotifications.map((notification) => ({
-    id: notification.id, // used for correctly handling cache in Apollo
-    networkId: notification.networkId, // used for filtering per network
-    timestamp: notification.created_at, // used for grouping / sorting
-    title: getMessageTitle(networks, notification), // title of notification
-    link: getPushLink(networks, notification), // link for click-through action
-    icon: getIcon(notification) // icon link
-  }))
+  const notifications = relevantNotifications
+    .filter(({ eventType }) => eventType !== eventTypes.VALIDATOR_ADDED)
+    .map((notification) => ({
+      id: notification.id, // used for correctly handling cache in Apollo
+      networkId: notification.networkId, // used for filtering per network
+      timestamp: notification.created_at, // used for grouping / sorting
+      title: getMessageTitle(networks, notification), // title of notification
+      link: getPushLink(networks, notification), // link for click-through action
+      icon: getIcon(notification) // icon link
+    }))
 
   return notifications
 }