+++ api/lib/schema.js
@@ -316,26 +316,10 @@ const typeDefs = gql`
     option: String
   }
 
-  type AccountInformation {
-    accountNumber: String
-    sequence: String
-  }
-
   type Powered {
     name: String
     providerAddress: String
     picture: String
-  }
-
-  type Overview {
-    networkId: String!
-    address: String!
-    totalStake: String!
-    totalStakeFiatValue: FiatValue
-    liquidStake: String!
-    totalRewards: String!
-    rewards: [Reward]
-    accountInformation: AccountInformation
   }
 
   enum EventType {
@@ -362,6 +346,10 @@ const typeDefs = gql`
 
   type Mutation {
     registerUser(idToken: String!): Boolean
+    notifications(
+      addressObjects: [NotificationInput]!
+      notificationType: String
+    ): Boolean
   }
 
   type ChainAppliedFees {
@@ -382,7 +370,7 @@ const typeDefs = gql`
   }
 
   type TransactionMetadata {
-    gasEstimate: Int!
+    gasEstimate: Int
     gasPrices: [GasPrice]
     chainAppliedFees: ChainAppliedFees
     accountSequence: Int
@@ -407,6 +395,12 @@ const typeDefs = gql`
   input NotificationInput {
     address: String!
     networkId: String!
+  }
+
+  input NotificationSetting {
+    topic: String!
+    type: String!
+    remove: Boolean
   }
 
   type Query {
@@ -442,11 +436,6 @@ const typeDefs = gql`
       denom: String!
       fiatCurrency: String
     ): Balance
-    overview(
-      networkId: String!
-      address: String!
-      fiatCurrency: String
-    ): Overview
     delegation(
       networkId: String!
       delegatorAddress: String!