+++ api/lib/source/cosmosV0-source.js
@@ -19,7 +19,6 @@ class CosmosV0API extends RESTDataSource {
     this.gasPrices = getNetworkGasPrices(network.id)
     this.store = store
     this.fiatValuesAPI = fiatValuesAPI
-    this.viewDenom = network.coinLookup[0].viewDenom
 
     this.setReducers()
   }
@@ -75,10 +74,6 @@ class CosmosV0API extends RESTDataSource {
     return this.getRetry(url)
   }
 
-  getStakingViewDenom() {
-    return this.viewDenom
-  }
-
   async getSignedBlockWindow() {
     const slashingParams = await this.query('/slashing/parameters')
     return slashingParams.signed_blocks_window
@@ -96,7 +91,7 @@ class CosmosV0API extends RESTDataSource {
     const txs = await this.loadPaginatedTxs(`txs?tx.height=${height}`)
     return Array.isArray(txs)
       ? this.reducers.transactionsReducerV2(
-          this.networkId,
+          this.network,
           txs,
           this.reducers,
           this.network.stakingDenom
@@ -247,7 +242,7 @@ class CosmosV0API extends RESTDataSource {
       return this.reducers.proposalReducer(
         this.network.id,
         proposal,
-        {}, //TODO also add tally to overview when we need it
+        {},
         totalBondedTokens
       )
     })
@@ -329,11 +324,15 @@ class CosmosV0API extends RESTDataSource {
     }
   }
 
-  async getBalancesFromAddress(address, fiatCurrency) {
+  // DEPRECATE
+  async getBalancesFromAddress(address, fiatCurrency, network) {
     this.checkAddress(address)
     const response = await this.query(`bank/balances/${address}`)
     let balances = response || []
-    const coins = balances.map(this.reducers.coinReducer)
+    const coins = balances.map((coin) => {
+      const coinLookup = network.getCoinLookup(network, coin.denom)
+      return this.reducers.coinReducer(coin, coinLookup)
+    })
     const fiatValues = await this.fiatValuesAPI.calculateFiatValues(
       coins,
       fiatCurrency
@@ -350,14 +349,17 @@ class CosmosV0API extends RESTDataSource {
     )
   }
 
-  async getBalancesV2FromAddress(address, fiatCurrency) {
+  async getBalancesV2FromAddress(address, fiatCurrency, network) {
     this.checkAddress(address)
     const [balancesResponse, delegations] = await Promise.all([
       this.query(`bank/balances/${address}`),
       this.getDelegationsForDelegatorAddress(address)
     ])
     const balances = balancesResponse || []
-    const coins = balances.map(this.reducers.coinReducer)
+    const coins = balances.map((coin) => {
+      const coinLookup = network.getCoinLookup(network, coin.denom)
+      return this.reducers.coinReducer(coin, coinLookup)
+    })
     const hasStakingDenom = coins.find(
       ({ denom }) => denom === this.network.stakingDenom
     )
@@ -507,25 +509,6 @@ class CosmosV0API extends RESTDataSource {
     )
     return uniqBy(allDelegations, 'delegator_address').map(
       ({ delegator_address }) => delegator_address
-    )
-  }
-
-  async getOverview(delegatorAddress, validatorsDictionary, fiatCurrency) {
-    this.checkAddress(delegatorAddress)
-    const [balances, delegations, undelegations] = await Promise.all([
-      this.getBalancesFromAddress(delegatorAddress),
-      this.getDelegationsForDelegatorAddress(delegatorAddress),
-      this.getUndelegationsForDelegatorAddress(delegatorAddress)
-    ])
-    const fiatValueAPI = this.fiatValuesAPI
-    return this.reducers.overviewReducer(
-      balances,
-      delegations,
-      undelegations,
-      this.network.stakingDenom,
-      fiatValueAPI,
-      fiatCurrency,
-      this.reducers
     )
   }