+++ app/src/App.vue
@@ -1,5 +1,5 @@
 <template>
-  <div id="app" :class="network">
+  <div id="app" :class="themeClass">
     <NetworkSelector />
     <AppHeader />
     <div id="app-content">
@@ -39,10 +39,15 @@ export default {
     NetworkSelector,
   },
   computed: {
-    ...mapState([`notifications`, `session`]),
+    ...mapState([`session`]),
     ...mapGetters([`network`]),
     isMobileApp() {
       return this.session.mobile
+    },
+    themeClass() {
+      return !this.$route.meta.networkSpecificRoute
+        ? `lunie-light`
+        : this.network
     },
   },
   store,