+++ app/src/components/network/NetworkItem.vue
@@ -1,5 +1,8 @@
 <template>
-  <div class="network-item" :class="{ disabled: disabled }">
+  <div
+    class="network-item"
+    :class="{ disabled: disabled }"
+  >
     <div class="network-icon">
       <img
         :src="`${networkItem.icon}`"
@@ -20,21 +23,21 @@
       hide-on-mobile
     />
     <div class="network-status">
+      <div v-if="isCurrentNetwork" class="network-selected">
+        <i class="material-icons notranslate">check</i>
+      </div>
       <img
-        v-if="network === networkItem.id"
+        v-else-if="disabled && isCurrentNetwork"
         class="tm-connected-network-loader"
         src="~assets/images/loader.svg"
         alt="a small spinning circle to display loading"
       />
-      <div v-else-if="!disabled && isCurrentNetwork" class="network-selected">
-        <i class="material-icons notranslate">check</i>
-      </div>
     </div>
   </div>
 </template>
 
 <script>
-import { mapGetters } from "vuex"
+import { mapGetters, mapState } from "vuex"
 import PoweredBy from "./PoweredBy"
 
 export default {
@@ -50,9 +53,10 @@ export default {
     disabled: {
       type: Boolean,
       default: false,
-    },
+    }
   },
   computed: {
+    ...mapState([`session`]),
     ...mapGetters([`network`]),
     isCurrentNetwork() {
       return this.networkItem.id === this.network