+++ app/tests/e2e/signin.spec.js
@@ -78,7 +78,6 @@ async function openMenu(browser) {
 }
 
 async function prepare(browser) {
-  browser.resizeWindow(400, 1024) // force mobile screen to be able to click some out of screen buttons
   await browser.url(
     browser.launch_url +
       browser.globals.slug +
@@ -112,6 +111,7 @@ async function prepare(browser) {
     },
     [browser.globals.network, browser.globals.address, browser.globals.wallet]
   )
+
   await browser.url(
     browser.launch_url +
       browser.globals.slug +
@@ -119,6 +119,7 @@ async function prepare(browser) {
   )
 
   // check if we are already signed in
+  await browser.click(".session-close")
   await browser.waitForElementVisible("#open-user-menu", 20000, true)
   await browser.click("#open-user-menu")
   await browser.waitForElementVisible("#create-new-account", 20000, true)