+++ app/CHANGELOG.md
@@ -1,11 +1,62 @@
 # Changelog
-
+ 
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 <!-- SIMSALA --> <!-- DON'T DELETE, used for automatic changelog updates -->
+
+## [1.0.220] - 2020-07-23
+
+### Added
+
+- Register clients for Email notifications @faboweb
+- [#4541](https://github.com/cosmos/lunie/issues/4541) Upgrade Polkadot js api to v1.25.1 @mariopino
+
+### Changed
+
+- [#4525](https://github.com/cosmos/lunie/issues/4525) CardSignIn is now a modal @Bitcoinera
+- [#4525](https://github.com/cosmos/lunie/issues/4525) Notifications page becomes landing page when user is signed in with email @Bitcoinera
+- [#4562](https://github.com/cosmos/lunie/pull/4562) Change close action in SessionFrame to make sense @Bitcoinera
+- [#4568](https://github.com/cosmos/lunie/pull/4568) Make block height in TmConnectedNetwork non-clickable if feature Activity is disabled @Bitcoinera
+- [#4525](https://github.com/cosmos/lunie/issues/4525) ux improvements for notifications page @jbibla
+
+### Fixed
+
+- [#4518](https://github.com/cosmos/lunie/issues/4518) Fix sending an expired idToken from localStorage to API (Apollo) @Bitcoinera
+- [#4534](https://github.com/cosmos/lunie/pull/4534) Fix the loading for notifications in PageNotifications bringing back the lunie SVG loader @Bitcoinera
+- [#4548](https://github.com/cosmos/lunie/pull/4548) Fix network item continuously displaying as loading @Bitcoinera
+- some route names were lowercase and some were uppercase and some router.push's were failing as a result @jbibla
+- router guard was sending user to an invalid route @jbibla
+
+## [1.0.219] - 2020-07-16
+
+### Added
+
+- [#4392](https://github.com/cosmos/lunie/issues/4392) Polkadot e2e tests @mariopino
+- [#4497](https://github.com/cosmos/lunie/issues/4497) Upgrade polkadot api to v1.24.1 @mariopino
+
+### Changed
+
+- [#3660](https://github.com/cosmos/lunie/issues/3660) Add the guide link to the memo field @Bitcoinera
+- cleaning up initial version of email auth and new user menu @jbibla
+
+### Fixed
+
+- [#4489](https://github.com/cosmos/lunie/issues/4489) Fix Paywall overflow and colors not using variables @Bitcoinera
+- Fix Android build @faboweb
+- Fix interface E2E tests @faboweb
+- Stored sessions had old network not the new one from the desired target network @faboweb
+- Connect to the first found network if desired network not found (fixes issues with local API) @faboweb
+
+### Deprecated
+
+- removed a bunch of unused components and associated tests @jbibla
+
+### Code Improvements
+
+- [#4042](https://github.com/cosmos/lunie/issues/4042) cleaning up TmPage part 1 @jbibla
 
 ## [1.0.218] - 2020-07-09